# src/integration-test/resources/OpenApiSpecificationIT/openapi.yaml
openapi: 3.0.1
info:
  title: api-contactsettings
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.0"
servers:
  - url: http://localhost:57894
    description: Generated server url
...
paths:
  /{municipalityId}/settings:
    get:
      tags:
        - ContactSettings
      summary: Find contact setting chain (including all delegates). Filter chain
        by the query-parameter.
      operationId: findByPartyIdAndQueryFilter
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: partyId
          in: query
          description: Party-ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
        - name: query
          in: query
          description: Filter query parameters. Only delegates that matches (i.e. has
            matching delegate filters) the specified query will be included in the result.
          required: true
          schema:
            $ref: "#/components/schemas/MultiValueMapStringString"
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
        - ContactSettings
      summary: Create contact setting
      operationId: create
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSettingCreateRequest"
        required: true
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/settings/{id}:
    get:
      tags:
        - ContactSettings
      summary: Read contact setting
      operationId: read
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Contact setting ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSetting"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ContactSetting"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
        - ContactSettings
      summary: Delete contact setting
      operationId: delete
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Contact setting ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
    patch:
      tags:
        - ContactSettings
      summary: Update contact setting
      operationId: update
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Contact setting ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSettingUpdateRequest"
        required: true
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSetting"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ContactSetting"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/settings/{id}/children:
    get:
      tags:
        - ContactSettings
      summary: Read virtual contact settings created by the instance idenitifed with
        the provided ID.
      operationId: readChildren
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: id
          in: path
          description: Contact setting ID
          required: true
          schema:
            type: string
          example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/settings/contact-channels:
    get:
      tags:
        - ContactSettings
      summary: Find contact setting filtered by contactChannel.destination
      operationId: findByDestination
      parameters:
        - name: municipalityId
          in: path
          description: Municipality id
          required: true
          schema:
            type: string
          example: 2281
        - name: destination
          in: query
          description: destination of contact channel
          required: true
          schema:
            type: string
          example: "0701234567"
      responses:
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSetting"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    ContactChannel:
      required:
        - alias
        - contactMethod
        - destination
      type: object
      properties:
        contactMethod:
          $ref: "#/components/schemas/ContactMethod"
        alias:
          type: string
          description: Alias for the destination
          example: Private phone
        destination:
          type: string
          description: Point of destination
          example: "+46701234567"
        disabled:
          type: boolean
          description: Signal if channel should be used or not when sending message
          example: true
          default: false
      description: ContactChannel model
    ContactMethod:
      type: string
      description: ContactMethod model
      enum:
        - SMS
        - EMAIL
    ContactSettingCreateRequest:
      required:
        - createdById
        - partyId
      type: object
      properties:
        partyId:
          type: string
          description: ID of the person or organization to whom the contact setting
            applies. Set to null when creating a 'virtual' contact setting.
          example: 15aee472-46ab-4f03-9605-68bd64ebc73f
        createdById:
          type: string
          description: ID of the contact setting that created this instance. Mandatory
            for virtual contact settings.
          example: 9ca9425e-42cf-4145-a9e7-d77e1ea9e5b0
        alias:
          type: string
          description: Alias for this contact setting
          example: My contact-settings
        contactChannels:
          type: array
          description: List of contact channels connected to this contact setting
          items:
            $ref: "#/components/schemas/ContactChannel"
      description: ContactSettingCreateRequest model
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    ContactSettingUpdateRequest:
      type: object
      properties:
        alias:
          type: string
          description: Alias for this contact setting
          example: My contact-setting
        contactChannels:
          type: array
          description: List of contact channels connected to this contact setting
          items:
            $ref: "#/components/schemas/ContactChannel"
      description: ContactSettingUpdateRequest model
    ContactSetting:
      type: object
      properties:
        id:
          type: string
          description: ID of the contact setting
          readOnly: true
          example: 0d64c132-3aea-11ec-8d3d-0242ac130003
        partyId:
          type: string
          description: ID of the person or organization to which the contact setting
            applies
          readOnly: true
          example: 15aee472-46ab-4f03-9605-68bd64ebc73f
        municipalityId:
          type: string
          description: Municipality ID
          readOnly: true
          example: "2281"
        createdById:
          type: string
          description: ID of the contact setting that created this instance. Applicable
            for virtual contact settings.
          readOnly: true
          example: 9ca9425e-42cf-4145-a9e7-d77e1ea9e5b0
        alias:
          type: string
          description: Alias for the person or organization to which the contact setting
            applies
          readOnly: true
          example: My contact-settings
        virtual:
          type: boolean
          description: Shows if the contact setting is virtual or not. A virtual instance
            doesn't have a partyId (i.e. doesn't have a direct relation to a real
            person/organization)
          readOnly: true
          example: false
        contactChannels:
          type: array
          description: List of contact channels connected to this contact setting
          readOnly: true
          items:
            $ref: "#/components/schemas/ContactChannel"
        created:
          type: string
          description: Timestamp when contact setting was created
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
        modified:
          type: string
          description: Timestamp when contact setting was last modified
          format: date-time
          readOnly: true
          example: 2020-08-31T01:30:00+02:00
      description: ContactSetting model
      readOnly: true
    MultiValueMapStringString:
      type: object
      properties:
        all:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        empty:
          type: boolean
      additionalProperties:
        type: array
        items:
          type: string
...
  securitySchemes: { }